73. Task 7 complete. Added new functions that searches/prints the population of the world. Also added similar tests for it.

+++ src/main/java/com/napier/sem/City.java
@@ -20,5 +20,5 @@ public class City
     /**
      * population
      */
-    public int population;
+    public double population;
 }
