+++ src/main/java/com/napier/sem/Country.java
@@ -25,7 +25,7 @@ public class Country
     /**
      * Population
      */
-    public int population;
+    public long population;
 
     /**
      * Capital
